Return status of a model

Checks for the presence of bbi_config.json file. If one exists, the run has "Finished Running". If a directory exists, but a config file has not been generated, the model is "Incomplete". If no directory exists, the model is "Not Run". This is 'source of truth' that all other model status helpers rely on.

bbi_nonmem_model_status(.mod)

# S3 method for bbi_model
bbi_nonmem_model_status(.mod)

# S3 method for bbi_nmboot_model
bbi_nonmem_model_status(.mod)

model_is_finished(.mod)

Arguments

.mod

a bbi_model object

Value

A character string. Either "Not Run", "Finished Running", or "Incomplete Run".

Logical (TRUE/FALSE)

Functions

model_is_finished(): Check if model run has finished (coerces the one of the three statuses to a logical value)