bbr

bbr is a library written in R that facilitates modeling and simulation by exposing functionality from the bbi project in R. From bbi's docs: bbi is a solution for managing projects involving modeling and simulation with a number of software solutions used in pharmaceutical sciences. This is a fork of the nonmemutils project that is broader in scope. Initial support encompasses NONMEM, however the API is designed in a way to be flexible to handle other software.

Details

bbr uses several fundamental objects, each of which has a number of S3 methods associated with it:

bbi_{.model_type}_model -- Represents a single model. Contains paths to any elements of this model that are stored on disk (e.g. a NONMEM control stream) as well as various fields containing metadata about the model. You can read more about these fields, and how to interact with them, in the following help pages: modify_tags() modify_notes() modify_based_on() modify_description() modify_bbi_args()

bbi_{.model_type}_summary -- Contains information and diagnostics about a model that has been run. See model_summary() for more information about contents and usage.

bbi_log_df -- A tibble containing information about multiple models. There are several sub-classes of bbi_log_df tibbles, variously containing metadata, information, and diagnostics about the models. See the following help pages for more information: run_log() summary_log() config_log()