Copy model from an existing NONMEM model

Private implementation function called by copy_model_from() dispatches. Create new .mod/ctl and new .yaml files based on a previous model. Used for iterating on model development. Also fills in necessary YAML fields for using run_log() later.

copy_nonmem_model_from(
  .parent_mod,
  .new_model,
  .description = NULL,
  .based_on_additional = NULL,
  .add_tags = NULL,
  .inherit_tags = FALSE,
  .update_model_file = TRUE,
  .overwrite = FALSE
)

Arguments

.parent_mod
S3 object of class bbi_nonmem_model to be used as the basis for copy.

.new_model
Path to the new model, either absolute or relative to the path to .parent_mod. Represents an absolute model path, which is the path to the YAML file and model file, both without extension, and the output directory (once the model is run). Numeric values will be coerced to character. See examples for usage.

.description
Character scalar description of new model run. This will be stored in the yaml (and can be viewed later in run_log()).

.based_on_additional
Character vector of path(s) to other models that this model was "based on." These are used to reconstuct model developement and ancestry. Paths must be relative to .new_model path. Note that the .parent_model will automatically be added to the based_on field, so no need to include that here.

.add_tags
Character vector with any new tags(s) to be added to {.new_model}.yaml

.inherit_tags
If FALSE, the default, new model will only have any tags passed to .add_tags argument. If TRUE inherit any tags from .parent_mod, with any tags passed to .add_tags appended.

.update_model_file
If TRUE, the default, update the $PROBLEM line in the new control stream. If FALSE, {.new_model}.[mod|ctl] will be an exact copy of its parent control stream.

.overwrite
If FALSE, the default, function will error if a model file already exists at specified .new_model path. If TRUE any existing file at .new_model will be overwritten silently.

Value

S3 object of class bbi_nonmem_model that can be passed to submit_nonmem_model()