Create bbi object

Constructor functions for objects of fundamental bbr classes.

create_model_object(res, save_yaml)

create_summary_object(res, .model_type = SUPPORTED_MOD_TYPES)

create_summary_list(res)

create_process_object(res)

create_run_log_object(log_df)

create_config_log_object(log_df)

create_summary_log_object(log_df)

create_log_df_impl(log_df, .class, .req_cols, .key)

Arguments

res

List to attempt to assign the class to

save_yaml

Logical scalar for whether to save the newly created model object to its corresponding YAML file and update the md5 hash.

.model_type

Character scalar of a valid model type (currently either nonmem or stan)

log_df

data.frame or tibble to attempt to assign the class to

.class

Character scalar of the class to assign

.req_cols

Character vector of required columns

.key

Character scalar of column name that will be treated as a primary key. Must be a type character, unique, and have no NA's or NULL's.

Functions

create_model_object(): Creates list object of class bbi_{.model_type}_model from named list with MODEL_REQ_INPUT_KEYS

create_summary_object(): Create list object of bbi_{.model_type}_summary class, first checking that all the required keys are present.

create_summary_list(): Create list output from model_summaries(). Each element will contain a bbi_{.model_type}_summary object, as well as some metadata about possible errors.

create_process_object(): Create list object of bbi_process class, first checking that all the required keys are present.

create_run_log_object(): Create tibble object of bbi_run_log_df class, first checking that all the required columns are present.

create_config_log_object(): Create tibble object of bbi_config_log_df class, first checking that all the required columns are present.

create_summary_log_object(): Create tibble object of bbi_summary_log_df class, first checking that all the required columns are present.

create_log_df_impl(): Implementation function used to create objects of class bbi_summary_log_df, bbi_config_log_df, and bbi_summary_log_df