Modify notes on a model object

Add, replace, or remove notes on a model object and corresponding YAML. Notes can be modified at any time (i.e. before or after a model is submitted).

add_notes(.mod, .notes)

replace_note(.mod, .old_note, .new_note)

replace_all_notes(.mod, .notes)

remove_notes(.mod, .notes)

Arguments

.mod
The bbi_{.model_type}_model object to modify

.notes
Character vector of notes to add or remove

.old_note
Character scalar of note to be replaced

.new_note
Character scalar of note that will be added

Value

The modified bbi_{.model_type}_model object

Details

The notes field on a bbi_{.model_type}_model object contains a character vector of brief notes about the model or modeling process.

These are often useful when looking at a bbi_run_log_df (the tibble output from run_log()); for example to remind yourself of decisions you made along the way, or of a particular thing you noticed when looking at model diagnostics.

Like tags, notes can also be collapsed using collapse_to_string() See the "Getting Started" vignette for an example of this.

Functions

add_notes: Add notes to a model object and corresponding YAML.

replace_note: Replaces a specific .old_note with .new_note on a model object and corresponding YAML. Warns and does nothing if .old_note is not present.

replace_all_notes: Replaces all notes on a model object and corresponding YAML with new notes.

remove_notes: Removes notes from a model object and corresponding YAML.

See also