Parses parameter estimates table

Returns a tibble containing parameter estimates from a bbi_{.model_type}_summary object. Details about the tibble that is returned are in the return value section below. Note: if you need to pull in the parameter estimates for a large number of NONMEM models at once, consider using param_estimates_batch() instead.

param_estimates(.summary, .alpha)

# S3 method for bbi_nonmem_summary
param_estimates(.summary, .alpha = NULL)

Arguments

.summary

A bbi_{.model_type}_summary object.

.alpha

numeric. If specified, return the p-value for each eta value on the diagonal, as well as whether they are below the specified value of alpha. Defaults to NULL.

Value

Returns a tibble with the following columns:

parameter_names -- Parameter name ("THETA1", "THETA2", etc.)

estimate -- Parameter estimate

stderr -- Standard Error

random_effect_sd -- OMEGA and SIGMA elements in standard deviation/correlation format

random_effect_sdse -- Standard errors to the OMEGA and SIGMA elements in standard deviation/correlation format

fixed -- TRUE if parameter is fixed, FALSE otherwise

diag -- TRUE if parameter is a diagonal element in random effect matrix, FALSE if off-diagonal, NA if parameter is a THETA

shrinkage -- Shrinkage for final estimation method, using SD parameterization (for NONMEM this is ETAshrinkSD and EPSshrinkSD). The shrinkage using Variance Parameterization, as well as for other estimation methods and multiple subpops if present, can be found in bbi_nonmem_summary$shrinkage_details.

Details

Note that Bayesian methods are not yet supported by this function. Creating a parameter table like this for Bayesian estimation methods requires a different approach, which has not yet been implemented. If the final estimation method of the input model is Bayesian, a not implemented error will be thrown.

Methods (by class)

param_estimates(bbi_nonmem_summary): Takes bbi_nonmem_summary object, the output of model_summary.bbi_nonmem_model().

See also