Batch Processing of Parameter Estimates

Calls out to bbi nonmem params and returns a tibble of parameter estimates. The primary reason for this function is for things like bootstraps or simulations where parameter estimates for a large number of models (hundreds or thousands) need to be pulled in together. In those cases, this function is much faster than using model_summaries() and param_estimates(), because it does not parse all of the other information contained in a bbi_nonmem_summary object.

param_estimates_batch(.path, ..., .dry_run = FALSE)

Arguments

.path

a path to a directory containing model sudirectories for batch parameter processing.

...

args passed through to bbi_exec()

.dry_run

show what the command would be without actually running it

Details

The tibble will always have columns absolute_model_path, run, error_msg, and termination_code. All other column names are those of the parameter estimates found in the .ext files detected in the directory passed. Note: the passed directory is not searched recursively. Passed directory will be searched for model files and output directories for found models will be searched for .ext files.

error_msg: column will be NA if error is not detected. Column will contain message if known error is detected. The known errors are

"no ext file contents": .ext file detected, but contains nothing

"no estimation output detected": .ext file detected, has contents, but is missing line -1000000000 which are the final parameter estimates

termination_code: parsed from line -1000000007 of the .ext

parameter columns: naming will be the same as names found in .ext file. If .ext files within the directory contain differing parameter names, the output tibble will show all parameter names across files and NA values for the files where a given parameter is missing.