Parse .bbi_args and return the three expected nmfe_options for NM-TRAN in the correct format

Parse .bbi_args and return the three expected nmfe_options for NM-TRAN in the correct format

parse_nmtran_args(.mod, .bbi_args = NULL, nmfe_options = NULL)

Arguments

.mod

A bbi_nonmem_model object.

.bbi_args

A named list specifying arguments to pass to NM-TRAN. Similar to the .bbi_args argument defined in submit_model(), though here only prdefault, tprdefault, and maxlim arguments are passed to NM-TRAN. nm_version is also supported and specifies which NM-TRAN executable to use. See print_bbi_args() for more details.

nmfe_options

named list of nmfe options defined in a bbi.yaml file.

Value

a named character vector

Details

Combines NONMEM submission args and consolidates to NMFE arguments only

The arguments of interest are prdefault, tprdefault, and maxlim, which impact the evaluation of NM-TRAN. Priority is .bbi_args > model yaml > bbi.yaml

Note: run_nmtran() considers only the first three NM-TRAN arguments (prdefault, tprdefault, and maxlim), but, starting with NONMEM 7.5, nmfe passes a fourth argument (do2test) to NM-TRAN.

bbi doesn't expose nmfe's -do2test argument, so in the bbi/bbr context this call always passes "0" to NM-TRAN for the do2test value. If that first call exits with a status of 9, nmfe does a follow-up call that drops the fourth argument (do2test).

Given the above, parse_nmtran_args() does not support the additional do2test argument for NONMEM 7.5. This should be revisited if run_nmtran() is called within submit_model() as a method of validating the control stream ahead of model execution. See additional discussion at https://github.com/metrumresearchgroup/bbr/pull/705.