Print methods for bbr objects

The various objects defined by bbr have their own print methods to allow users to get a quick view of the contents of the object. When printing a bbi object in an .Rmd file that is intended to be knit, consider setting results = 'asis' in the chunk options. This will make for prettier formatting, especially of table outputs.

# S3 method for class 'bbi_process'
print(x, ..., .call_limit = 250)

# S3 method for class 'bbi_model'
print(x, ...)

# S3 method for class 'bbi_nonmem_summary'
print(x, .digits = 3, .fixed = FALSE, .off_diag = FALSE, .nrow = NULL, ...)

# S3 method for class 'bbi_nmboot_summary'
print(x, .digits = 3, .nrow = 10, ...)

# S3 method for class 'nmtran_process'
print(x, ...)

# S3 method for class 'model_tree_static'
print(x, newpage = is.null(vp), vp = NULL, ...)

Arguments

x

plot to display

...

other arguments not used by this method

.call_limit

Integer scalar for the max number of characters to print before truncating the call string. This is compared with the entire length, but only the positional arguments between the executable path and the first long option will be truncated.

.digits

Number of significant digits to use for parameter table. Defaults to 3.

.fixed

If FALSE, the default, omits fixed parameters from the parameter table.

.off_diag

If FALSE, the default, omits off-diagonals of OMEGA and SIGMA matrices from the parameter table.

.nrow

If NULL, the default, print all rows of the parameter table. If 0, don't print table at all. Otherwise, prints only .nrow rows.

newpage

Logical (T/F). If TRUE, draw new (empty) page first.

vp

viewport to draw plot in

Functions

print(bbi_process): Prints the call made to bbi and whether the process is still running or has finished.

print(bbi_model): Prints the information contained in the model object and whether the model has been run

print(bbi_nonmem_summary): Prints a high level summary of a model from a bbi_nonmem_summary object

print(bbi_nmboot_summary): Prints a high level summary of a model from a bbi_nmboot_summary object

print(nmtran_process): Prints the NM-TRAN evaluation of a bbi_nonmem_model object

print(model_tree_static): Draw model tree as a static plot

Examples

if (FALSE) { # \dontrun{
pl_tree <- model_tree(MODEL_DIR, static = TRUE)

print(pl_tree)
print(pl_tree, vp = grid::viewport(width=0.5, height=0.5))
print(pl_tree, newpage = TRUE, vp = grid::viewport(width=0.5, height=0.5))
} # }