Add data from model_summaries() to bbi_run_log_df

Runs model_summaries() on all models in the input and return a subset of the resulting summaries as a tibble. summary_log() will return a new tibble with the "absolute_model_path" column as the primary key, and all the columns extracted from model_summaries() (but none of the other columns from the input tibble). add_summary() returns the input tibble, with all the columns extracted from model_summaries() joined onto it. See details section for fields included.

summary_log(.mods, ..., .keep_bbi_object = FALSE)

add_summary(.log_df, ..., .keep_bbi_object = FALSE)

Arguments

.mods
The model object that will be passed through to model_summaries() before constructing the output tibble. Could be a bbi_run_log_df tibble, a list of bbi_{.model_type}_model objects, a character vector of file paths to models, a numeric vector of integers corresponding to a file names of a models.

...
Arguments passed through to model_summaries().

.keep_bbi_object
FALSE by default. If TRUE, a list column will be added containing the full bbi_nonmem_summary object for each row. Use this if you would like to extract additional fields from the summary object.

.log_df
a bbi_run_log_df tibble

Details

The following fields from bbi_nonmem_summary are extracted and included by default. If you would like more fields, you can set .keep_bbi_object = TRUE and extract them manually.

error_msg -- Error message from model_summary(). If NULL the call succeeded. If not NULL, the rest of the fields will be NULL.

needed_fail_flags -- Logical for whether the call initially failed, but passed with the inclusion of .fail_flags

bbi_summary -- The full bbi_nonmem_summary object. This will only be included if .keep_bbi_object = TRUE was passed.

ofv -- Objective function value from last estimation method, using $ofv$ofv_no_constant field.

param_count -- Count of (non-fixed) parameters estimated in final estimation method.

estimation_method -- Character vector of estimation method(s) used. Extracted from $run_details

problem_text -- Character vector of text from $PROB. Extracted from $run_details

number_of_patients -- Extracted from $run_details

number_of_obs -- Extracted from $run_details

covariance_step_aborted -- Extracted from $run_heuristics

large_condition_number -- Extracted from $run_heuristics

correlations_not_ok -- Extracted from $run_heuristics

parameter_near_boundary -- Extracted from $run_heuristics

hessian_reset -- Extracted from $run_heuristics

has_final_zero_gradient -- Extracted from $run_heuristics

minimization_terminated -- Extracted from $run_heuristics

Functions

add_summary: Create summary_log() tibble and join against the input bbi_run_log_df tibble.