Verify YAML file integrity

These functions use an md5 digest of the model YAML file, stored each time a model is loaded into memory, to keep track of any changes made to the YAML that are not reflected in the object held in memory.

reconcile_yaml(.mod)

check_yaml_in_sync(.mod)

Arguments

.mod
bbi_{.model_type}_model object

Functions

reconcile_yaml: Use to manually reconcile model object in memory with its YAML file. Extracts YAML path from model object and pulls in YAML file. Any shared keys are overwritten with the values from the YAML and new keys in YAML are added to the model object. The md5 digest is then updated to reflect the new state.

check_yaml_in_sync: Checks that model YAML file is the same as when it was last read into the model object. Errors if the md5 digests are not identical. This is called internally in most functions that interact with a model object.