Create a tibble from model_summaries() data

Runs model_summaries() on all models in the input and returns a subset of the each resulting summary as a tibble.

summary_log(.base_dir, .recurse = FALSE, .include = NULL, ...)

add_summary(.log_df, ...)

Arguments

.base_dir

Base directory to look in for models.

.recurse

If FALSE, the default, only include models in .base_dir but not subdirectories. This is passed through to fs::dir_ls() – If TRUE recurse fully, if a positive number the number of levels to recurse.

.include

A character vector specifying which runs or model tags to include in the run log.

...

Arguments passed through to model_summaries()

.log_df

a bbi_run_log_df tibble (the output of run_log())

Value

An object of class bbi_summary_log_df, which includes the fields described below. If all model summaries fail, the returned tibble will only contain the absolute_model_path, run, and error_msg columns.

summary_log() creates a new tibble with one row per model found in .base_dir (and subdirectories, if .recurse = TRUE).

add_summary() adds these fields to the tibble passed to .log_df.

Details

The following fields from bbi_nonmem_summary (the output of model_summary()) are extracted and included by default. If you would like more fields from the summary object, you can extract them manually from the bbi_summary list column.

error_msg – Error message from model_summary(). If NULL the call succeeded. If not NULL, the rest of the fields will be NULL.

needed_fail_flags – Logical for whether the call initially failed, but passed with the inclusion of .fail_flags. See model_summaries() docs for more details.

bbi_summary – The full bbi_nonmem_summary object for each row. This can be queried further by extracting it as a list, or by using dplyr::mutate() etc.

ofv – Objective function value with no constant from the final estimation method. The constant, and the value with the constant can be found in $ofv.

aic, bic – Akaike information criterion and Bayesian information criterion. These are calculated using NONMEM's objective function value with constant (available as of NONMEM 7.4) as approximations of the -2 log-likelihood.

NA is reported for models where the final estimation method is one for which these derived quantities are not meaningful for model comparison (e.g., SAEM or Bayesian methods).

param_count – Count of (non-fixed) parameters estimated in final estimation method.

estimation_method – Character vector of estimation method(s) used. Extracted from $run_details.

problem_text – Character vector of text from $PROB. Extracted from $run_details.

number_of_subjects – Count of unique subjects in the input data set, extracted from $run_details.

number_of_obs – Total count of observations in the input data set, extracted from $run_details.

condition_number – The condition number for the final estimation method, if present.

any_heuristics – Logical indicating whether any of the columns extracted from $run_heuristics are TRUE. Duplicative information, but helpful for filtering.

$run_heuristics columns – One logical column for each element extracted from $run_heuristics. These are named and described in the model_summary() docs.

See also